Support SprT file present in Slithereens
